tr2/music_main: refactor music track handling
This refactors the way music is checked when requested to play tracks.
Some tracks require to be enforced such as from flip effects, while
others are tracked, such as from regular triggers. Tracked tracks will
not play in succession, provided the mode is the same when requesting
to play them.

Resolves #1899.