tr1/data: fix mesh positions in Egyptian levels
This fixes the angle of the drawbridge in Obelisk of Khamoon when open
so that Lara's shadow isn't occluded and the artefacts aren't
embedded. It also lifts the senet table and lowers the toppled chair in
room 57. The toppled chair is also fixed in Return to Egypt and Temple
of the Cat.

Resolves #2006.