libtrx/rooms/common: append commands to existing triggers
This will handle cases in old levels where a sector has multiple
trigger entries. Officially only one is supported, but it seems old
levels use this approach to add dummy triggers for such things as
bridges on top of pre-existing triggers. This would have worked with
the old g_TriggerIndex approach because of the way floor height tests
were performed. Instead now, if we detect a trigger but the sector
already has one, we simply append the new commands onto the existing
trigger.

Resolves #2114.