gun: make common Lara gun flash routine
This removes the duplication in TR2 for adding dynamic light from
Lara's guns, and gets TR1 to use the same routine, which takes into
consideration the config option. The logic for showing the shotgun
gunflare has been adjusted to allow for the two options to work
independently of each other.